Move cli logging to 'log' module
